Fix merge conflicts

+++ crates/primitives/src/traits/block/body.rs
@@ -68,17 +68,17 @@ pub trait BlockBody:
 
     /// Returns whether or not the block body contains any blob transactions.
     fn has_blob_transactions(&self) -> bool {
-        self.transactions().iter().any(|tx| tx.ty() as u8 == TxType::Eip4844 as u8)
+        self.transactions().iter().any(|tx| tx.ty() == TxType::Eip4844 as u8)
     }
 
     /// Returns whether or not the block body contains any EIP-7702 transactions.
     fn has_eip7702_transactions(&self) -> bool {
-        self.transactions().iter().any(|tx| tx.ty() as u8 == TxType::Eip7702 as u8)
+        self.transactions().iter().any(|tx| tx.ty() == TxType::Eip7702 as u8)
     }
 
     /// Returns an iterator over all blob transactions of the block
     fn blob_transactions_iter(&self) -> impl Iterator<Item = &Self::SignedTransaction> + '_ {
-        self.transactions().iter().filter(|tx| tx.ty() as u8 == TxType::Eip4844 as u8)
+        self.transactions().iter().filter(|tx| tx.ty() == TxType::Eip4844 as u8)
     }
 
     /// Returns only the blob transactions, if any, from the block body.
